Dyamics update
Adds an instance of Will's Dynamic Processor to compute a gain expansion factor for the internally OSmpld negative feedback path.

This is first lowpassed and so simulates some of the transformer highpassing effect.

Finally brings some life to the system. Positive FB also provided for for dynamic bass boosting. Effect is massively interactive with the various gain and biasing levels.

Expansion ratio set to 1:5 for now with knee 6db. Attack 5ms adjustable release.

There seemed to be an error in the original soft knee equation. See fxobject.h