Sagging etc.
Adds sag controls after Kuehnel's Fender Bassman analysis. Controls to input/output volume, diode Vt and resistance.

Experimentally determined a linear relationship between Vt and the maximum clipped output level. Supplies that inverse gain makeup to the input as well so the effect can be heard for it's knee hardness and breakup differences in morphing from transistor to tube like.

Found and applied, a good exponential fit to level the saturation current.

Used the custom messaging system to provide an efficient means of visual feedback for the dynamic parameters.

Added option to test between double/float precision for the BJT and Diodes.